+++ source/usbloader/apploader.h
@@ -10,7 +10,7 @@ extern "C"
 typedef void (*entry_point)(void);
 
 /* Prototypes */
-s32 Apploader_Run(entry_point *, u8, u8, u8, u8, u8, u8, u8);
+s32 Apploader_Run(entry_point *, u8, u8, u8, u8, u8, u8, u32);
 void gamepatches(void * dst, int len, u8 videoSelected, u8 patchcountrystring, u8 vipatch);
 
 #ifdef __cplusplus
